Remove all makefiles in make distclean
Change the build system to descend into every directory where we
create a Makefile, but not to build or run anything during "make all"
and "make check" in directories we previously didn't visit.  Document
specific build targets that can be used in those directories.

Do not generate a Makefile for the securid_sam2 module unless we are
building it, for consistency with other conditionally built
directories.