Remove unnecessary directories
Remove the plugin modules wpse, cksum_body, and locate/python, which
aren't used by the test suite or built by default.

Remove util/collected-client-lib, as we no longer have a need to
create a smaller client-only library.

Remove util/gss-kernel-lib, as it turned out not to be useful for
facilitating kernel integrations.