Add bttest unlink page command
To enable testing of recursive btree traversal, add an unlink page
command to the bttest program (used for debugging the libdb2 btree
back end).  This new bttest command can unlink a specified page
number, or it can search for and unlink a page that has both a left
and a right sibling.  (The user can specify whether to find an
internal page or a leaf page.)

This unlinking makes the page inaccessible to conventional sequential
traversal, simulating some btree corruption that has been seen in the
field.

ticket: 8476