Fix KDC lookaside cache to correctly track entry count
The KDC lookaside cache was incrementing num_entries when an
entry was added, but it did not decrement num_entries when an
entry was removed.  The num_entries count is decremented in
discard_entry.  Also, fix some function comments to correctly
explain what the function is doing.  Refactor kdc_insert_lookaside
by moving the code that inserts an entry into a new static
function called insert_entry.