Slightly rework the k5srvutil man page
Do not give the impression that arbitrary new keys can be added
to the keytab [1], since only the new keys randomly generated by the
KDC via 'k5srvutil change' can be added to the keytab.

Reiterate the importance of running k5srvutil delold after
running k5srvutil change in the description of k5srvutil change,
as well as in the description of k5srvutil delold itself.

[1] Requested by Daniel Kahn Gillmor <dkg@fifthhorseman.net>