KCM: Fix detection of a subsidiary cache
A KCM cache collection is typically denoted as:
    KCM:USER_ID
while a subsidiary cache inside a collection is normally denoted as:
    KCM:USER_ID:CACHE_ID
Where a cache ID is an internal identification of the credentials cache
the KCM deamon tracks.

The previous detection of a subsidiary cache in KCM's kcm_ptcursor_new()
treated any defname with name longer than 4 characters as a subsidiary cache.

Instead, let's check if there are more color (:) characters in the cache
location string.