Add code to properly handle plugins for dlclose()
dlclose() for plugins can be called in 2 cases:
- when libgssapi_krb5.so is itself dlclose()ed, in which case the library
destructor will fire and the freeMechList() function will go through the
list of mechanisms and explicitly call dlclose() on each mechanism.
- when exit() is called in a binary linking to libgssapi_krb5.so, in which
case plugins may be dlclose()ed before the libgssapi_krb5.so destructor
is called.

In the second case the freeMechList() function will try to close the plugins
again. On some libcs with certain linker flags used (ex. RTLD_NODELETE) this
may cause an assert() to be thrown.

To avoid calling dlclose() twice we introduce a destroyer hook, which should
be called by plugins when their destroyer is called (this implicitly requires
plugins to implement a destroyer).
When the destroyer hook is called by a plugin a no_close flag is set so no
further attempts to dlclose() the plugin are made.

If a plugin does not provide a function to register the hook, we assume we
can't properly handle dlclose() on these plugins and prevent any dlclose()
from happening. This may cause some leaks when libgssapi_krb5.so is explicitly
dlclose()ed, but that is the lesser evil.

Signed-off-by: Simo Sorce <simo@redhat.com>