Avoid increased alignment restriction warnings
In kdb_log.h, cast through void * after computing the address in the
INDEX macro.

In ipropd_svc.c, use a void * instead of a char * as the generic
handler return value.

In rc4.c, cast through void * when using the cipher state data pointer
as a structure pointer.

In sha256.c and sha512.c, cast through void * when using the save
buffer as a structure pointer.  (This code may not be conformant, but
it should work in practice given the offsets of the save field in the
sha256state and sha512state structures.)

[ghudson@mit.edu: rewrote commit message]