Update Travis build for warning-clean clang build
Create a custom build matrix which passes -Werror to the clang build
via a make variable.  (Using a configure variable does not currently
work, as some of our configure test programs generate warnings.)

Also set the language to C++ (so we use clang++ for the C++ test
programs and not g++ when compiling with clang), and turn on the
maintainer-mode checks for the Travis build.