Consistently name and constify address variables
In libkdb5, libapputils, the KDC, kadmind, and both KDB modules, use
the name "remote_addr" for the variable containing the remote address.
In schpw.c:process_chpw_request(), use the name "local_addr" for the
parameter containing the local address.  Make the remote_addr
parameter const in libkdb5 and the DAL.

[ghudson@mit.edu: combined commits and rewrote commit message]