Stop building CNS for Windows
CNS is an old Windows ticket manager which was long ago supplanted by
leash.  We have been building it and copying it in "nmake install",
but not including it in the KfW installer.  Remove it from the build,
and trim windows/lib down to just the part used by leashdll and leash
(loadfuncs.obj).

Also stop generating a Makefile for windows/wintel; it was removed
from the build in commit 530043377ee6f39b9ca95c13f6260602e6fe7674.

ticket: 8682 (new)