Clean up Windows build CPU detection
In the main build, set a variable BITS based on the detected CPU type,
and use it in library name expansions rather than repeatedly checking
for the same three 64-bit architectures.  Also remove an unused
variable WLIB, formerly used to name the winsock library.

In the installer build, detect the CPU type in the makefile and pass
it to candle with the -arch option.  This sets $(sys.BUILDARCH) in wix
scripts and also sets the default value of Win64 in components so we
don't have to set it for each one.

Update the build instructions and appveyor.yml to reflect that the
builder no longer needs to set CPU.

ticket: 8662