Use PTHREAD_CFLAGS when testing for getpwnam_r()
On Solaris, AX_PTHREAD can include -D_POSIX_PTHREAD_SEMANTICS in
PTHREAD_CFLAGS, which affects which variants of getpwnam_r() and
getpwuid_r() are declared by system headers.  In configure.in, use
PTHREAD_CFLAGS when testing for those functions to make sure that the
tests see the same variants as the build will.

[ghudson@mit.edu: rewrote commit message and comment]

ticket: 8723 (new)