Fix memory leak in gss_add_cred() creation case
If gss_add_cred() is called with no input_cred_handle, it creates a
new credential with one element.  At the end of the function, use the
created credential as the output container, rather than creating a
second one and leaking the first.

Add a test program for gss_add_cred() and run it.

ticket: 8729
tags: pullup
target_version: 1.16-next
target_version: 1.15-next