zap.so is required when unit tests are not compiled with clang or gcc
k5-platform.h does not pick up memset() for zap() when unit tests
are being built by Solaris/developer studio. In this case we
must link with zap.so explicitly.

not sure if proposed change in this pull request is what we
also want to have for gcc & clang

New commit incorporates Greg Hudson's suggestion:
	 s/zap.so/zap.o
it works.