Simplify negState choice in SPNEGO initiator
In the SPNEGO initiator code, simplify the choice of the outgoing
negState value, and remember the acceptor negState choice throughout
the process.  The outgoing negState value is REJECT if sending an
error token, ACCEPT_COMPLETE when sending a final MIC, and
ACCEPT_INCOMPLETE otherwise.

RFC 4178 permits negState to be omitted in some cases, so rename
ACCEPT_DEFECTIVE_TOKEN to UNSPECIFIED for clarity.  Use this value as
the acceptor negState for the first pass through
spnego_gss_init_sec_context() when there is no acceptor token.

Decide whether to return GSS_S_COMPLETE or GSS_S_CONTINUE_NEEDED at
the end of processing, instead of it being a shared responsibility of
the helper functions.  Return GSS_S_COMPLETE on success in the helper
functions and use "goto cleanup" in a few more places.  Leave
handle_mic() alone as it is also used by the acceptor code.